+++ Moder.Core/Helpers/AppTheme.cs
@@ -1,0 +1,8 @@
+using Avalonia.Styling;
+
+namespace Moder.Core.Helpers;
+
+public static class AppTheme
+{
+    
+}
