+++ Moder.Core/Models/ThemeMode.cs
@@ -1,8 +1,8 @@
-namespace Moder.Core.Resources;
+namespace Moder.Core.Models;
 
 public enum ThemeMode : byte
 {
     Light,
     Dark,
-    DarkSlateGray,
+    DarkSlateGray
 }
