+++ Moder.Core/Services/GameResources/Modifiers/ModifierService.cs
@@ -98,6 +98,7 @@ public sealed class ModifierService
 
     private static ModifierEffectType GetModifierType(string modifierName, string modifierFormat)
     {
+        // TODO: 重新支持从数据库中定义修饰符
         // if (_modifierTypes.TryGetValue(modifierName, out var modifierType))
         // {
         //     return modifierType;
