+++ Moder.Core/ViewsModel/Menus/AppInitializeControlViewModel.cs
@@ -4,7 +4,7 @@ using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 using Moder.Core.Infrastructure;
 using Moder.Core.Messages;
-using Moder.Core.Resources;
+using Moder.Core.Models;
 using Moder.Core.Services;
 using Moder.Core.Services.Config;
 using Moder.Language.Strings;
